+++ js/app/services/albumartservice.ts
@@ -5,7 +5,7 @@
  * later. See the COPYING file.
  *
  * @author Pauli Järvinen <pauli.jarvinen@gmail.com>
- * @copyright 2024 Pauli Järvinen
+ * @copyright 2024 - 2026 Pauli Järvinen
  *
  */
 
@@ -50,6 +50,8 @@ ng.module('Music').service('albumartService', [function() {
 					setPlaceholder(element, art.name, art.artist?.name + art.name);
 				} else if (art.title) {
 					setPlaceholder(element, art.title);
+				} else {
+					setPlaceholder(element, '?');
 				}
 			}
 		}